Fixed status streaming and batch testing

+++ notary-eth-integration-test/src/integration-test/kotlin/integration/eth/DepositIntegrationTest.kt
@@ -68,7 +68,7 @@ class DepositIntegrationTest {
             integrationHelper.waitOneIrohaBlock()
 
             Assertions.assertEquals(
-                BigDecimal(amount, ETH_PRECISION.toInt()).add(BigDecimal(initialAmount)),
+                BigDecimal(amount, ETH_PRECISION).add(BigDecimal(initialAmount)),
                 BigDecimal(integrationHelper.getIrohaAccountBalance(clientIrohaAccountId, etherAssetId))
             )
         }
@@ -103,7 +103,7 @@ class DepositIntegrationTest {
             integrationHelper.waitOneIrohaBlock()
 
             Assertions.assertEquals(
-                BigDecimal(amount, ETH_PRECISION.toInt()).add(BigDecimal(initialAmount)),
+                BigDecimal(amount, ETH_PRECISION).add(BigDecimal(initialAmount)),
                 BigDecimal(integrationHelper.getIrohaAccountBalance(clientIrohaAccountId, etherAssetId))
             )
         }
